Project health: fix error: comparison of integers of different signs.
	Change on 2019/05/22 by antoniocortes <antoniocortes@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=249461128

+++ protobuf/runtime/src/com/google/protobuf/RepeatedField.m
@@ -140,7 +140,7 @@ void CGPRepeatedMessageFieldRemove(CGPRepeatedField *field, jint index) {
   uint32_t count = CGPRepeatedFieldSize(field);
   id *msgBuffer = (id *)field->data->buffer;
   [msgBuffer[index] autorelease];
-  if (count > index + 1) {
+  if (count > (uint32_t)index + 1) {
     memmove(&(msgBuffer[index]),
             &(msgBuffer[index + 1]),
             sizeof(id) * (count - (index + 1)));
